+++ package.json
@@ -12,6 +12,8 @@
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-config-htmlacademy": "0.5.2",
+    "moment": "2.24.0",
+    "moment-locales-webpack-plugin": "1.2.0",
     "webpack": "4.41.2",
     "webpack-cli": "3.3.10",
     "webpack-dev-server": "3.9.0"
