+++ src/main.js
@@ -4,7 +4,7 @@ import FilmsSectionComponent from './components/film-section/film-section';
 import FilmsStatisticsComponent from './components/film-statistics/films-statistics';
 import PageController from './controllers/page-controller';
 import { Films, RenderPosition } from './consts';
-import { render } from './utils/generator';
+import { render } from './utils/render';
 
 import { generateFilm } from './mock/film';
 import { generateProfile } from './mock/profile';
